requirements.txt
requests==2.32.3
python-dotenv
python-telegram-bot==21.1
nest_asyncio
Flask==3.1.0
blinker==1.9.0
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.8
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
urllib3==2.2.2
Werkzeug==3.1.3
